Add HierarchyCreator.createHierarchy static method, which returns fully initialized Hierarchy object, instead of depending on lazy initialization
